+++ Chac/Chac/Source/Domain/Interfaces/GeocodingService.swift
@@ -1,0 +1,12 @@
+//
+//  GeocodingService.swift
+//  Chac
+//
+//  Created by 가은 on 2/3/26.
+//
+
+import CoreLocation
+
+protocol GeocodingService {
+    func fetchLocationName(from location: CLLocation) async -> String
+}
